% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoBoot.R
\name{cocoBoot}
\alias{cocoBoot}
\title{Bootstrap Confidence Intervals for Autocorrelations of a COCO Model}
\usage{
cocoBoot(
  coco,
  numb.lags = 21,
  rep.Bootstrap = 400,
  confidence = 0.95,
  plot_main = "Bootstrap",
  xlab = "Lag",
  ylab = "Autocorrelation",
  julia = FALSE
)
}
\arguments{
\item{coco}{An object of class coco.fit or coco.fit.c}

\item{numb.lags}{Number of lags for which to compute autocorrelations}

\item{rep.Bootstrap}{Number of bootstrap replicates to use}

\item{confidence}{Confidence level for the intervals}

\item{plot_main}{Plot title}

\item{xlab}{X-axis label for the plot}

\item{ylab}{Y-axis label for the plot}
}
\value{
A plot of the autocorrelations with bootstrap confidence intervals
}
\description{
Computes bootstrap confidence intervals for the autocorrelations of a COCO model. The function can handle both Poisson and GP models, with order 1 or 2. The function also has options for the number of lags, the number of bootstrap replicates, the confidence level, and the labels for the plot.
}
\references{
Tsay, R. S. (1992) Model checking via parametric bootstraps in time series analysis. \emph{Applied Statistics} \bold{41}, 1--15.
}
