% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoSoc.R
\name{cocoSoc}
\alias{cocoSoc}
\title{Compute Scores for Various Models}
\usage{
cocoSoc(data, models = "all", print.progress = TRUE, max_x_score = 50, ...)
}
\arguments{
\item{data}{A numeric vector containing the data to be used for modeling.}

\item{models}{A character string specifying which models to use. Default is `"all"`, which uses both Poisson and GP models.}

\item{print.progress}{A logical value indicating whether to print progress messages. Default is `TRUE`.}

\item{max_x_score}{An integer which is used as the maximum count for the computation
of the score. The default value is `50`}

\item{...}{Additional arguments to be passed to the `cocoReg` function.}
}
\value{
A list of class `"cocoVarsoc"` containing:
\describe{
  \item{fits}{A list of fitted model objects.}
  \item{scores_list}{A list of score objects for each model.}
  \item{scores_df}{A data frame containing the logarithmic, quadratic, and ranked probability scores for each model.}
}
}
\description{
This function computes and returns scores for Poisson and Generalized Poisson models.
}
