% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoVarsoc.R
\name{cocoVarsoc}
\alias{cocoVarsoc}
\title{Compute Scores for Various Models}
\usage{
cocoVarsoc(
  data,
  models = "all",
  print.progress = TRUE,
  val.num.score = 1e-10,
  ...
)
}
\arguments{
\item{data}{A numeric vector containing the data to be used for modeling.}

\item{models}{A character string specifying which models to use. Default is `"all"`, which uses both Poisson and GP models.}

\item{print.progress}{A logical value indicating whether to print progress messages. Default is `TRUE`.}

\item{val.num.score}{A numeric value specifying the validation number score. Default is `1e-10`.}

\item{...}{Additional arguments to be passed to the `cocoReg` function.}
}
\value{
A list of class `"cocoVarsoc"` containing:
\describe{
  \item{fits}{A list of fitted model objects.}
  \item{scores_list}{A list of score objects for each model.}
  \item{scores_df}{A data frame containing the logarithmic, quadratic, and ranked probability scores for each model.}
}
}
\description{
This function computes and returns scores for Poisson and Generalized Poisson models. It resembles Stata's `varsoc` function.
}
