\name{cuts}
\alias{cuts}
\title{
  Time series of monthly counts of claimants collecting wage loss benefit for injuries in
the workplace
}
\description{
  Monthly counts of claimants collecting wage loss benefit for injuries in
the workplace at one specific service delivery location of the Workers Compensation Board
of British Columbia, Canada in the period January 1985 to December 1994. Only injuries due to cuts and lacerations are considered. The data have been provided by Brendan McCabe. 
}

\keyword{Data}