% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGP1.R
\name{dGP1}
\alias{dGP1}
\title{Compute the probability density function (PDF) of GP1 convolution Distribution}
\usage{
dGP1(x, y, par)
}
\arguments{
\item{x}{The value at which to evaluate the PDF}

\item{y}{non-negative integer}

\item{par}{A vector of 3 parameters: lambda, alpha and eta}
}
\value{
The PDF of the GP1 convolution distribution at the given value
}
\description{
Computes the PDF of the GP1 convolution distribution with parameters lambda, alpha and eta, at a given value x, for a given value of y .
}
