% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dGP2.R
\name{dGP2}
\alias{dGP2}
\title{Compute the probability density function (PDF) of the GP2 convolution.}
\usage{
dGP2(x, y, z, par)
}
\arguments{
\item{x}{The value at which to evaluate the PDF}

\item{y}{non-negative integer}

\item{z}{non-negative integer}

\item{par}{A vector of 5 parameters: lambda, alpha1, alpha2, alpha3 and eta}
}
\value{
The PDF of the the GP2 convolution distribution at the given value
}
\description{
Computes the PDF of the GP2 convolution. with parameters lambda,alpha1, alpha2, alpha3 and eta, at a given value x, for a given value of y and z .
}
