\name{goldparticle}
\alias{goldparticle}
\title{
  Time Series of Gold particles Counts in a well-efined Colloidal
Solution
}
\description{
  A sample of 370 counts of gold particles in a well-defined colloidal
solution at equidistant points in time, originally published
in Westgren (1916) and used in Jung and Tremanye (2006).  
}

\source{
R.C. Jung, A.R. Tremayne (2006) Coherent forecasting in integer time series models. 
  \emph{International Journal of Forecasting} \bold{22}, 223--238
  
Westgren, A. (1916). Die Veraenderungsgeschwindigkeit der lokalen
Teilchenkonzentration in kollioden Systemen (Erste Mitteilung).
\emph{Arkiv foer Matematik, Astronomi och Fysik}, \bold{11}, 1--24.
}


\keyword{Data}