% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGP1.R
\name{pGP1}
\alias{pGP1}
\title{Compute the cumulative distribution function (CDF) of the GP1 convolution}
\usage{
pGP1(q, y, par)
}
\arguments{
\item{q}{Value at which to evaluate the CDF}

\item{y}{non-negative integer}

\item{par}{A vector of 3 parameters: lambda, alpha and eta}
}
\value{
The CDF of the the GP1 convolution distribution at the given quantile
}
\description{
Computes the CDF of the the GP2 convolution distribution with parameters lambda, alpha and eta, at a given quantile q, for a given value of y .
}
