% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGP2.R
\name{pGP2}
\alias{pGP2}
\title{Compute the cumulative distribution function (CDF) of the random operator used in the GP2 model}
\usage{
pGP2(q, y, z, par)
}
\arguments{
\item{q}{The value at which to evaluate the CDF}

\item{y}{non-negative integer}

\item{z}{non-negative integer}

\item{par}{A vector of 5 parameters: lambda, alpha1, alpha2, alpha3 and eta}
}
\value{
The CDF of the the GP2 convolution distribution at the given quantile
}
\description{
Computes the CDF of the GP2 convolution distribution with parameters lambda,alpha1, alpha2, alpha3 and eta, at a given value q, for a given value of y and z .
}
