% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.coco.R
\name{print.summary.coco}
\alias{print.summary.coco}
\title{print.summary.coco}
\usage{
\method{print}{summary.coco}(coco, score = FALSE, julia = FALSE, ...)
}
\arguments{
\item{coco}{an object of class cocoReg}

\item{score}{logical indicating whether to include score statistics in the summary}

\item{julia}{if TRUE, scores are computed via the julia implementation}
}
\value{
summary of the cocoReg analysis
}
\description{
Summarizes the results of a cocoReg analysis
}
\author{
Manuel Huth
}
