% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgenpois.R
\name{rgenpois}
\alias{rgenpois}
\title{Generate Random Numbers from Generalized Poisson Distribution}
\usage{
rgenpois(n, lambda, eta)
}
\arguments{
\item{n}{Number of random numbers to generate}

\item{lambda}{Parameter for the generalized Poisson distribution}

\item{eta}{Parameter for the generalized Poisson distribution}
}
\value{
A vector of random numbers from the generalized Poisson distribution
}
\description{
Generates a vector of random numbers from a generalized Poisson distribution with parameters lambda and eta. The function uses a while loop to generate the numbers.
}
