% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setJuliaSeed.R
\name{setJuliaSeed}
\alias{setJuliaSeed}
\title{Set Seed for Julia's Random Number Generator}
\usage{
setJuliaSeed(julia_seed)
}
\arguments{
\item{julia_seed}{An integer seed value to be passed to Julia's random number generator.}
}
\description{
Sets the seed for Julia's random number generator to ensure reproducibility.
}
\details{
This function initializes the necessary Julia functions and sets the random seed for Julia. 
If the provided seed is NULL, the function does nothing.
}
\author{
Your Name (or the appropriate author's name)
}
