% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoBoot.R
\name{cocoBoot}
\alias{cocoBoot}
\title{Bootstrap Confidence Intervals for Autocorrelations of a COCO Model}
\usage{
cocoBoot(
  coco,
  numb.lags = 21,
  rep.Bootstrap = 400,
  confidence = 0.95,
  plot_bootstrap = FALSE,
  plot_main = "Bootstrap",
  xlab = "Lag",
  ylab = "Autocorrelation",
  julia = FALSE,
  julia_seed = NULL
)
}
\arguments{
\item{coco}{An object of class coco.fit or coco.fit.c}

\item{numb.lags}{Number of lags for which to compute autocorrelations}

\item{rep.Bootstrap}{Number of bootstrap replicates to use}

\item{confidence}{Confidence level for the intervals}

\item{plot_bootstrap}{if TRUE, a graph with the results is plotted}

\item{plot_main}{Plot title}

\item{xlab}{X-axis label for the plot}

\item{ylab}{Y-axis label for the plot}

\item{julia}{if TRUE, the bootstrap is run with Julia.}

\item{julia_seed}{Seed for the julia implementation. Only used if julia equals TRUE.}
}
\value{
A plot of the autocorrelations with bootstrap confidence intervals
}
\description{
Model checking procedure emphasizing reproducibility in fitted models to provide an overall evaluation of fit as proposed by Tsay (1992).
}
\details{
Computes bootstrap confidence intervals for the autocorrelations of a fitted model and compares them to the sample autocorrelations of the data.
}
\examples{
lambda <- 1
alpha <- 0.4
set.seed(12345)
data <- cocoSim(order = 1, type = "Poisson", par = c(lambda, alpha), length = 100)$data
fit <- cocoReg(order = 1, type = "Poisson", data = data)

#assessment using bootstrap - R implementation
#boot_r <- cocoBoot(fit, rep.Bootstrap=50)
}
\references{
Tsay, R. S. (1992) Model checking via parametric bootstraps in time series analysis. \emph{Applied Statistics} \bold{41}, 1--15.
}
