% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoForecast.R
\name{cocoForecast}
\alias{cocoForecast}
\title{Forecast for COCO models}
\usage{
cocoForecast(
  coco,
  max = 15,
  xcast = NULL,
  plot = TRUE,
  title = "Probability mass",
  xlab = "x",
  ylab = "Probabilities",
  width_bars = 0.04,
  seasonality = c(1, 2),
  decimals = 4,
  julia = FALSE
)
}
\arguments{
\item{coco}{An object of class coco.fit or coco.fit.c}

\item{max}{The maximum number of forecast steps}

\item{xcast}{A vector of covariate values for forecasting (only required for coco.fit.c class)}

\item{plot}{A logical value indicating whether to plot the forecast}

\item{title}{Plot title}

\item{xlab}{X-axis label for the plot}

\item{ylab}{Y-axis label for the plot}

\item{width_bars}{Width of bars in the plot}

\item{seasonality}{A vector of two integers indicating the seasonalities of the time series}

\item{decimals}{Number of decimal places for the forecast probabilities}

\item{julia}{if TRUE, the estimate is predicted with Julia.}
}
\value{
A list containing the plot, density, mode, and median of the forecast
}
\description{
Computes forecast for a COCO model. The function can handle both Poisson and GP models, with order 1 or 2. The function also has options for the maximum number of forecast steps, covariates for forecasting, the plot, plot title, axis labels, width of bars, seasonality, and decimal precision.
}
