% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoPit.R
\name{cocoPit}
\alias{cocoPit}
\title{Probability Integral Transform Based Model Assessment Procedure}
\usage{
cocoPit(coco, J = 10, alpha = 0.05, julia = FALSE)
}
\arguments{
\item{coco}{An object of class coco}

\item{J}{Number of bins for the histogram (default: 10)}

\item{alpha}{Confidence level for the chi square test of equal distributions.}

\item{julia}{if TRUE, the PIT is computed with Julia.}
}
\value{
The probability integral transform and p-values for the coco object.
}
\description{
Computes the probability integral transform (PIT) and provides the non-randomized PIT histogram for assessing absolute performance of a fitted model as proposed by Czado et al. (2009).
}
\details{
Details need to be added
}
\examples{
lambda <- 1
alpha <- 0.4
set.seed(12345)
data <- cocoSim(order = 1, type = "Poisson", par = c(lambda, alpha), length = 100)$data
#julia_installed = TRUE ensures that the fit object
#is compatible with the julia cocoPit implementation 
fit <- cocoReg(order = 1, type = "Poisson", data = data)

#PIT R implementation
pit_r <- cocoPit(fit)
}
\references{
Czado, C., Gneiting, T. and Held, L. (2009) Predictive model assessment for count data. \emph{Biometrics} \bold{65}, 1254--61.

Jung, R. C. and Tremayne, A. R. (2011) Convolution-closed models for count time series with applications. \emph{Journal of Time Series Analysis}, \bold{32}, 3, 268--280.
}
\author{
Manuel Huth
}
