% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoPit.R
\name{cocoPit}
\alias{cocoPit}
\title{Probability integral transform plot for coco}
\usage{
cocoPit(
  coco,
  J = 10,
  ylab = "Relative frequency",
  xlab = "Probability integral transform",
  plot_main = "PIT",
  julia = FALSE
)
}
\arguments{
\item{coco}{An object of class coco.fit or coco.fit.c}

\item{J}{Number of bins for the histogram (default: 10)}

\item{ylab}{Label for the y-axis (default: "Relative frequency")}

\item{xlab}{Label for the x-axis (default: "Probability integral transform")}

\item{plot_main}{Title for the plot (default: "PIT")}

\item{julia}{if TRUE, the PIT is computed with Julia.}
}
\value{
A plot of the probability integral transform for the coco object.
}
\description{
Plots the probability integral transform (PIT) for a given coco object.
}
\examples{
lambda <- 1
alpha <- 0.4
set.seed(12345)
data <- cocoSim(order = 1, type = "Poisson", par = c(lambda, alpha), length = 100)$data
#julia_installed = TRUE ensures that the fit object
#is compatible with the julia cocoPit implementation 
fit <- cocoReg(order = 1, type = "Poisson", data = data, julia_installed = TRUE)

#PIT R implementation
pit_r <- cocoPit(fit)
#PIT Julia implementation
pit_julia <- cocoPit(fit, julia = TRUE)
}
\references{
Czado, C., Gneiting, T. and Held, L. (2009) Predictive model assessment for count data. \emph{Biometrics} \bold{65}, 1254--61.

Jung, R. C. and Tremayne, A. R. (2010) Convolution-closed models for count time series with applications. \emph{Journal of Time Series Analysis}, \bold{32}, 3, 268--280.
}
\author{
Manuel Huth
}
