% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoResid.R
\name{cocoResid}
\alias{cocoResid}
\title{cocoResid}
\usage{
cocoResid(coco, val.num = 1e-05)
}
\arguments{
\item{coco}{An object of class "coco.fit" or "coco.fit.c"}

\item{val.num}{A non-negative real number which is used to stop the calculation of}
}
\value{
A list with the elements:

  - fitted: a vector of the fitted values
  
  - varX: a vector of the conditional variances of the residuals
}
\description{
The function calculates the residuals of a coco model object.
}
\author{
Manuel Huth
}
