% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuts.R
\docType{data}
\name{cuts}
\alias{cuts}
\title{Time Series of Monthly Counts of Claimants Collecting Wage Loss Benefit}
\format{
A time series (`ts`) object containing monthly data from January 1985 to December 1994.
}
\source{
Freeland, R. K.; McCabe, B.P.M. (2004) 
  Analysis of count data by means of the Poisson autoregressive model. 
  \emph{Journal of Time Series Analysis}, \bold{25}, 701--722.
}
\usage{
cuts
}
\description{
Monthly counts of claimants collecting wage loss benefit for injuries in the workplace 
at one specific service delivery location of the Workers Compensation Board 
of British Columbia, Canada in the period January 1985 to December 1994. 
Only injuries due to cuts and lacerations are considered. 
The data have been provided by Brendan McCabe.
}
\examples{
plot(cuts)

}
\keyword{datasets}
