% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\docType{data}
\name{downloads}
\alias{downloads}
\title{Time Series of Daily Downloads of a TeX-Editor}
\format{
A time series (\code{ts}) object containing daily data from June 2006 to February 2007.
}
\source{
Weiss, C.H. (2008)
Thinning operations for modelling time series of counts -- a survey.
\emph{Advances in Statistical Analysis}, \bold{92}, 319--341.
}
\usage{
downloads
}
\description{
The data represent the number of daily downloads of a TeX-editor between
June 2006 and February 2007. The dataset contains 267 observations.
The data have been provided by Christian Weiss.
}
\keyword{datasets}
