\name{downloads}
\alias{downloads}
\title{
  Time Series of Daily Downloads of a TeX-Editor
}
\description{
  The data represent the number of daily downloads of a TeX-editor between June 2006 and February
2007 and has a sample size of 267. The data have been provided by Christian Weiss. 
}


\keyword{Data}