\name{goldparticle}
\alias{goldparticle}
\title{
  Time Series of Gold particles Counts in a well-efined Colloidal
Solution
}
\description{
  A sample of 370 counts of gold particles in a well-defined colloidal
solution at equidistant points in time, originally published
in Westgren (1916) and used in Jung and Tremanye (2006).  
}

\source{
  source

  references to?  
}


\keyword{Data}