% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installJuliaPackages.R
\name{installJuliaPackages}
\alias{installJuliaPackages}
\title{installJuliaPackages}
\usage{
installJuliaPackages()
}
\value{
no return value, called to install Julia packages in Julia.
}
\description{
checks for needed Julia packages and installs them if not installed.
}
